fix(admin): address webhook review findings
