+++ admin-ui/app/api/proxy/__tests__/route.test.ts
@@ -22,68 +22,25 @@ const mockAppendProxyHeaders = vi.fn((_req: unknown, headers: Headers) => {
   }
 });
 const mockGetRequestBody = vi.fn(() => Promise.resolve(undefined));
-const mockBuildProxyResponse = vi.fn(async (response: Response) => {
-  // Return a lightweight object that mimics NextResponse enough for assertions
-  const body = await response.text().catch(() => '');
-  return {
-    __proxy: true,
-    body,
-    status: response.status,
-    headers: new Headers(response.headers),
-  };
-});
 
 vi.mock('@/lib/api-config', () => ({
   buildApiUrlForRequest: mockBuildApiUrlForRequest,
 }));
 
-vi.mock('@/lib/server-auth', () => ({
-  getBackendAuthHeaders: mockGetBackendAuthHeaders,
-  appendProxyHeaders: mockAppendProxyHeaders,
-  getRequestBody: mockGetRequestBody,
-  buildProxyResponse: mockBuildProxyResponse,
-}));
+vi.mock('@/lib/server-auth', async (importOriginal) => {
+  const actual = await importOriginal<typeof import('@/lib/server-auth')>();
+  return {
+    ...actual,
+    getBackendAuthHeaders: mockGetBackendAuthHeaders,
+    appendProxyHeaders: mockAppendProxyHeaders,
+    getRequestBody: mockGetRequestBody,
+  };
+});
 
 vi.mock('@/lib/logger', () => ({
   logger: { error: vi.fn(), warn: vi.fn(), info: vi.fn(), debug: vi.fn() },
 }));
 
-// We need NextResponse.json for the error branches in the route
-vi.mock('next/server', () => {
-  class MockNextRequest {
-    method: string;
-    url: string;
-    headers: Headers;
-    cookies: { get: (name: string) => { value: string } | undefined };
-    nextUrl: { pathname: string; search: string };
-
-    constructor(url: string, init?: { method?: string; headers?: HeadersInit }) {
-      this.url = url;
-      this.method = init?.method ?? 'GET';
-      this.headers = new Headers(init?.headers);
-      this.cookies = { get: () => undefined };
-      const parsed = new URL(url);
-      this.nextUrl = {
-        pathname: parsed.pathname,
-        search: parsed.search,
-      };
-    }
-  }
-
-  return {
-    NextRequest: MockNextRequest,
-    NextResponse: {
-      json: (body: unknown, init?: { status?: number; headers?: HeadersInit }) => ({
-        __errorResponse: true,
-        body,
-        status: init?.status ?? 200,
-        headers: new Headers(init?.headers),
-      }),
-    },
-  };
-});
-
-// Import NextRequest from our mock so we can construct requests
 import { NextRequest } from 'next/server';
 
 // ---------------------------------------------------------------------------
@@ -129,10 +86,8 @@ describe('Proxy route – forward()', () => {
     const { GET } = await import('@/app/api/proxy/[...path]/route');
     const result = await GET(makeRequest('/some/path'));
 
-    // buildProxyResponse should have been called with the backend response
-    expect(mockBuildProxyResponse).toHaveBeenCalledTimes(1);
-    // The result comes from our mock buildProxyResponse
-    expect(result).toMatchObject({ __proxy: true, status: 200 });
+    expect(result.status).toBe(200);
+    await expect(result.json()).resolves.toEqual({ data: 'hello' });
   });
 
   // 2. Auth header injection
@@ -170,6 +125,8 @@ describe('Proxy route – forward()', () => {
         'cache-control': 'no-store',
         pragma: 'no-cache',
         'x-request-id': 'backend-request-id',
+        'content-length': '9',
+        'set-cookie': 'backend-session=do-not-forward',
       },
     }));
     vi.stubGlobal('fetch', fetchMock);
@@ -185,11 +142,13 @@ describe('Proxy route – forward()', () => {
     expect(forwarded.get('authorization')).toBe('Bearer cookie-token');
     expect(forwarded.get('if-match')).toBe('"admin-webhook-41-r1"');
     expect(forwarded.get('idempotency-key')).toBe('0123456789abcdef');
-    const responseHeaders = (result as unknown as { headers: Headers }).headers;
+    const responseHeaders = result.headers;
     expect(responseHeaders.get('etag')).toBe('"admin-webhook-41-r2"');
     expect(responseHeaders.get('cache-control')).toBe('no-store');
     expect(responseHeaders.get('pragma')).toBe('no-cache');
     expect(responseHeaders.get('x-request-id')).toBe('test-request-id');
+    expect(responseHeaders.get('content-length')).toBeNull();
+    expect(responseHeaders.get('set-cookie')).toBeNull();
   });
 
   it('does not reflect or log webhook request canaries from a backend 422', async () => {
@@ -228,19 +187,15 @@ describe('Proxy route – forward()', () => {
     }));
 
     const serialized = JSON.stringify({
-      body: (result as unknown as { body: string }).body,
-      headers: Object.fromEntries(
-        (result as unknown as { headers: Headers }).headers.entries(),
-      ),
+      body: await result.clone().text(),
+      headers: Object.fromEntries(result.headers.entries()),
     });
     for (const canary of Object.values(canaries)) {
       expect(serialized).not.toContain(canary);
     }
-    expect((result as unknown as { headers: Headers }).headers.get('cache-control')).toBe('no-store');
-    expect((result as unknown as { headers: Headers }).headers.get('pragma')).toBe('no-cache');
-    expect((result as unknown as { headers: Headers }).headers.get('x-request-id')).toBe(
-      'test-request-id',
-    );
+    expect(result.headers.get('cache-control')).toBe('no-store');
+    expect(result.headers.get('pragma')).toBe('no-cache');
+    expect(result.headers.get('x-request-id')).toBe('test-request-id');
     expect(logger.error).not.toHaveBeenCalled();
   });
 
@@ -251,7 +206,7 @@ describe('Proxy route – forward()', () => {
     const { GET } = await import('@/app/api/proxy/[...path]/route');
     const result = await GET(makeRequest('/data'));
 
-    expect((result as unknown as { headers: Headers }).headers.get('x-request-id')).toBe('test-request-id');
+    expect(result.headers.get('x-request-id')).toBe('test-request-id');
   });
 
   // 3c. x-request-id is set on error response
@@ -261,7 +216,8 @@ describe('Proxy route – forward()', () => {
     const { POST } = await import('@/app/api/proxy/[...path]/route');
     const result = await POST(makeRequest('/data', 'POST'));
 
-    expect(result).toMatchObject({ __errorResponse: true, status: 502 });
+    expect(result.status).toBe(502);
+    await expect(result.json()).resolves.toEqual({ detail: 'Backend unavailable' });
   });
 
   // 4. Timeout returns 504
@@ -297,11 +253,8 @@ describe('Proxy route – forward()', () => {
 
     const result = await resultPromise;
 
-    expect(result).toMatchObject({
-      __errorResponse: true,
-      body: { detail: 'Backend request timed out' },
-      status: 504,
-    });
+    expect(result.status).toBe(504);
+    await expect(result.json()).resolves.toEqual({ detail: 'Backend request timed out' });
   });
 
   // 5. Network error returns 502
@@ -314,11 +267,8 @@ describe('Proxy route – forward()', () => {
     const { POST } = await import('@/app/api/proxy/[...path]/route');
     const result = await POST(makeRequest('/data', 'POST'));
 
-    expect(result).toMatchObject({
-      __errorResponse: true,
-      body: { detail: 'Backend unavailable' },
-      status: 502,
-    });
+    expect(result.status).toBe(502);
+    await expect(result.json()).resolves.toEqual({ detail: 'Backend unavailable' });
   });
 
   // 6. GET retry on network error
@@ -341,7 +291,8 @@ describe('Proxy route – forward()', () => {
     const result = await resultPromise;
 
     expect(fetchMock).toHaveBeenCalledTimes(2);
-    expect(result).toMatchObject({ __proxy: true, status: 200 });
+    expect(result.status).toBe(200);
+    await expect(result.json()).resolves.toEqual({ retried: true });
   });
 
   // 7. POST does NOT retry
@@ -356,10 +307,8 @@ describe('Proxy route – forward()', () => {
 
     // Only one call — no retry for POST
     expect(fetchMock).toHaveBeenCalledTimes(1);
-    expect(result).toMatchObject({
-      __errorResponse: true,
-      status: 502,
-    });
+    expect(result.status).toBe(502);
+    await expect(result.json()).resolves.toEqual({ detail: 'Backend unavailable' });
   });
 
   // 8. Request body forwarding
@@ -426,11 +375,8 @@ describe('Proxy route – forward()', () => {
 
     // Two attempts for GET
     expect(fetchMock).toHaveBeenCalledTimes(2);
-    expect(result).toMatchObject({
-      __errorResponse: true,
-      body: { detail: 'Backend unavailable' },
-      status: 502,
-    });
+    expect(result.status).toBe(502);
+    await expect(result.json()).resolves.toEqual({ detail: 'Backend unavailable' });
   });
 
   // 12. AbortError on GET does NOT trigger retry
@@ -444,10 +390,7 @@ describe('Proxy route – forward()', () => {
 
     // Only one call — AbortError should NOT trigger retry
     expect(fetchMock).toHaveBeenCalledTimes(1);
-    expect(result).toMatchObject({
-      __errorResponse: true,
-      body: { detail: 'Backend request timed out' },
-      status: 504,
-    });
+    expect(result.status).toBe(504);
+    await expect(result.json()).resolves.toEqual({ detail: 'Backend request timed out' });
   });
 });
