+++ admin-ui/app/webhooks/page.tsx
@@ -1,7 +1,6 @@
 'use client';
 
-import { Fragment, useCallback, useEffect, useRef, useState } from 'react';
-import { flushSync } from 'react-dom';
+import { Fragment } from 'react';
 import {
   AlertTriangle,
   ChevronLeft,
@@ -43,114 +42,14 @@ import {
   TableHeader,
   TableRow,
 } from '@/components/ui/table';
-import { usePrivilegedActionDialog } from '@/components/ui/privileged-action-dialog';
-import { useToast } from '@/components/ui/toast';
-import {
-  canonicalWebhookApi,
-  detectWebhookApi,
-  legacyWebhookApi,
-} from '@/lib/api-client';
-import type { LegacyWebhookDeliveryView, LegacyWebhookView } from '@/lib/api-client';
 import { formatDateTime } from '@/lib/format';
+import type { WebhookStatus } from '@/types';
 import {
-  WebhookApiError,
-  WebhookContractError,
-  WebhookTransportError,
-} from '@/lib/http';
-import {
-  createIdempotentCommand,
-} from '@/lib/idempotent-command';
-import type { IdempotentCommand } from '@/lib/idempotent-command';
-import type {
-  WebhookCatalog,
-  WebhookCreateRequest,
-  WebhookListResponse,
-  WebhookPatchRequest,
-  WebhookRegistration,
-  WebhookSecretResponse,
-  WebhookStatus,
-} from '@/types';
-
-const PAGE_SIZE = 20;
-
-type WebhookMode = 'canonical' | 'legacy' | null;
-type SecretOperation = 'create' | 'rotate';
-
-type SecretCommandResult = {
-  data: WebhookSecretResponse;
-  etag: string;
-  status: number;
-  requestId: string | null;
-};
-
-type PendingSecretCommand = {
-  command: IdempotentCommand<SecretCommandResult>;
-  operation: SecretOperation;
-  webhookId: number | null;
-};
-
-type SecretState = {
-  value: string;
-  replayed: boolean;
-  operation: SecretOperation;
-};
-
-type EditorState = {
-  kind: 'metadata' | 'destination';
-  registration: WebhookRegistration;
-};
-
-type ConflictState = {
-  status: 412 | 428;
-  action: string;
-  registration: WebhookRegistration;
-};
-
-type SafeError = {
-  message: string;
-  requestId: string | null;
-};
-
-const canLoadCanonicalData = (status: WebhookStatus): boolean => (
-  status.mode !== 'off'
-  && status.schema_ready
-  && status.migration.phase === 'complete'
-);
-
-const safeError = (error: unknown, fallback: string): SafeError => {
-  if (
-    error instanceof WebhookApiError
-    || error instanceof WebhookContractError
-    || error instanceof WebhookTransportError
-  ) {
-    return { message: error.message, requestId: error.requestId };
-  }
-  return { message: fallback, requestId: null };
-};
-
-const isConditionalError = (error: unknown): error is WebhookApiError & { status: 412 | 428 } => (
-  error instanceof WebhookApiError && (error.status === 412 || error.status === 428)
-);
-
-const activationBlockReason = (
-  registration: WebhookRegistration,
-  status: WebhookStatus,
-): string | null => {
-  if (registration.active) return null;
-  if (!status.delivery_capability_ready) return 'Delivery capability is unavailable';
-  if (registration.secret_rotation_required) return 'Generate a new signing secret before activation';
-  if (status.key_state !== 'available') return 'Webhook signing key is unavailable';
-  if (status.limits.active_registrations_over_limit) return 'Active registration limit is exceeded';
-  return null;
-};
-
-const registrationReviewSummary = (registration: WebhookRegistration): string => (
-  `Webhook ${registration.id}, revision ${registration.revision}: `
-  + `${registration.target_display}; ${registration.active ? 'active' : 'inactive'}; `
-  + `${registration.timeout_seconds}s timeout; `
-  + `description "${registration.description || 'None'}"; `
-  + `events ${registration.event_types.join(', ')}.`
-);
+  activationBlockReason,
+  canLoadCanonicalData,
+  useWebhooksPageController,
+  WEBHOOK_PAGE_SIZE,
+} from './use-webhooks-page-controller';
 
 function StatusAlerts({ status }: { status: WebhookStatus }) {
   const required = status.migration.secret_rotation_required_count;
@@ -225,613 +124,83 @@ function StatusAlerts({ status }: { status: WebhookStatus }) {
 }
 
 function WebhooksPageContent() {
-  const promptPrivileged = usePrivilegedActionDialog();
-  const { success, error: showError } = useToast();
-
-  const [mode, setMode] = useState<WebhookMode>(null);
-  const [status, setStatus] = useState<WebhookStatus | null>(null);
-  const [catalog, setCatalog] = useState<WebhookCatalog | null>(null);
-  const [canonicalPage, setCanonicalPage] = useState<WebhookListResponse>({
-    items: [],
-    total: 0,
-    limit: PAGE_SIZE,
-    offset: 0,
-  });
-  const [legacyItems, setLegacyItems] = useState<LegacyWebhookView[]>([]);
-  const [legacyTotal, setLegacyTotal] = useState(0);
-  const [offset, setOffset] = useState(0);
-  const [loading, setLoading] = useState(true);
-  const [statusError, setStatusError] = useState<SafeError | null>(null);
-
-  const [createOpen, setCreateOpen] = useState(false);
-  const [createUrl, setCreateUrl] = useState('');
-  const [createDescription, setCreateDescription] = useState('');
-  const [createTimeout, setCreateTimeout] = useState('10');
-  const [createEvents, setCreateEvents] = useState<string[]>([]);
-  const [legacyEvents, setLegacyEvents] = useState('');
-  const [legacyEnabled, setLegacyEnabled] = useState(true);
-  const [creating, setCreating] = useState(false);
-
-  const [editor, setEditor] = useState<EditorState | null>(null);
-  const [editDescription, setEditDescription] = useState('');
-  const [editTimeout, setEditTimeout] = useState('10');
-  const [editEvents, setEditEvents] = useState<string[]>([]);
-  const [replacementUrl, setReplacementUrl] = useState('');
-  const [mutatingId, setMutatingId] = useState<number | null>(null);
-  const [conflict, setConflict] = useState<ConflictState | null>(null);
-
-  const [secretState, setSecretState] = useState<SecretState | null>(null);
-  const [secretCopied, setSecretCopied] = useState(false);
-  const [secretAcknowledged, setSecretAcknowledged] = useState(false);
-  const [secretWarning, setSecretWarning] = useState('');
-  const [commandError, setCommandError] = useState('');
-  const [commandBusy, setCommandBusy] = useState(false);
-  const [pendingOperation, setPendingOperation] = useState<SecretOperation | null>(null);
-  const pendingCommandRef = useRef<PendingSecretCommand | null>(null);
-  const secretRef = useRef<SecretState | null>(null);
-
-  const [legacyExpandedId, setLegacyExpandedId] = useState<string | null>(null);
-  const [legacyDeliveries, setLegacyDeliveries] = useState<LegacyWebhookDeliveryView[]>([]);
-  const [legacyDeliveryLoading, setLegacyDeliveryLoading] = useState(false);
-
-  const clearCreateForm = useCallback(() => {
-    setCreateUrl('');
-    setCreateDescription('');
-    setCreateTimeout('10');
-    setCreateEvents([]);
-    setLegacyEvents('');
-    setLegacyEnabled(true);
-  }, []);
-
-  const clearSensitiveCommandState = useCallback((synchronous = false) => {
-    pendingCommandRef.current = null;
-    secretRef.current = null;
-    const clearState = () => {
-      setSecretState(null);
-      setSecretCopied(false);
-      setSecretAcknowledged(false);
-      setSecretWarning('');
-      setCommandError('');
-      setCommandBusy(false);
-      setPendingOperation(null);
-    };
-    if (synchronous) {
-      flushSync(clearState);
-      return;
-    }
-    clearState();
-  }, []);
-
-  const loadControlPlane = useCallback(async (requestedOffset = 0) => {
-    setLoading(true);
-    setStatusError(null);
-    try {
-      const detected = await detectWebhookApi();
-      setStatus(detected.status);
-      setMode(detected.kind);
-      setConflict(null);
-
-      if (detected.kind === 'canonical') {
-        setLegacyItems([]);
-        setLegacyTotal(0);
-        setLegacyExpandedId(null);
-        setLegacyDeliveries([]);
-        if (!canLoadCanonicalData(detected.status)) {
-          setCatalog(null);
-          setCanonicalPage({ items: [], total: 0, limit: PAGE_SIZE, offset: 0 });
-          setOffset(0);
-          return;
-        }
-        const [nextCatalog, nextPage] = await Promise.all([
-          detected.client.getWebhookCatalog(),
-          detected.client.getWebhooks({ limit: PAGE_SIZE, offset: requestedOffset }),
-        ]);
-        setCatalog(nextCatalog);
-        setCanonicalPage(nextPage);
-        setOffset(nextPage.offset);
-        return;
-      }
-
-      setCatalog(null);
-      setCanonicalPage({ items: [], total: 0, limit: PAGE_SIZE, offset: 0 });
-      const legacyPage = await detected.client.getWebhooks({
-        limit: PAGE_SIZE,
-        offset: requestedOffset,
-      });
-      setLegacyItems(legacyPage.items);
-      setLegacyTotal(legacyPage.total);
-      setOffset(requestedOffset);
-    } catch (error) {
-      setMode(null);
-      setStatus(null);
-      setCatalog(null);
-      setCanonicalPage({ items: [], total: 0, limit: PAGE_SIZE, offset: 0 });
-      setLegacyItems([]);
-      setLegacyTotal(0);
-      setStatusError(safeError(error, 'Webhook status could not be loaded.'));
-    } finally {
-      setLoading(false);
-    }
-  }, []);
-
-  const loadCanonicalPage = useCallback(async (requestedOffset: number) => {
-    setLoading(true);
-    try {
-      const nextPage = await canonicalWebhookApi.getWebhooks({
-        limit: PAGE_SIZE,
-        offset: requestedOffset,
-      });
-      setCanonicalPage(nextPage);
-      setOffset(nextPage.offset);
-    } catch (error) {
-      const bounded = safeError(error, 'Webhook registrations could not be loaded.');
-      showError('Unable to load webhooks', bounded.message);
-    } finally {
-      setLoading(false);
-    }
-  }, [showError]);
-
-  useEffect(() => {
-    void loadControlPlane(0);
-  }, [loadControlPlane]);
-
-  useEffect(() => {
-    const handleBeforeUnload = (event: BeforeUnloadEvent) => {
-      if (!secretRef.current && !pendingCommandRef.current) return;
-      event.preventDefault();
-      event.returnValue = '';
-    };
-    const handlePageHide = () => {
-      clearSensitiveCommandState(true);
-    };
-    const handlePageShow = (event: PageTransitionEvent) => {
-      if (event.persisted) clearSensitiveCommandState(true);
-    };
-    window.addEventListener('beforeunload', handleBeforeUnload);
-    window.addEventListener('pagehide', handlePageHide);
-    window.addEventListener('pageshow', handlePageShow);
-    return () => {
-      window.removeEventListener('beforeunload', handleBeforeUnload);
-      window.removeEventListener('pagehide', handlePageHide);
-      window.removeEventListener('pageshow', handlePageShow);
-      pendingCommandRef.current = null;
-      secretRef.current = null;
-    };
-  }, [clearSensitiveCommandState]);
-
-  const recoverConditionalConflict = useCallback(async (
-    error: WebhookApiError & { status: 412 | 428 },
-    webhookId: number,
-    action: string,
-  ) => {
-    try {
-      const current = await canonicalWebhookApi.getWebhook(webhookId);
-      setCanonicalPage((page) => ({
-        ...page,
-        items: page.items.map((registration) => (
-          registration.id === webhookId ? current.data : registration
-        )),
-      }));
-      setConflict({ status: error.status, action, registration: current.data });
-    } catch {
-      setConflict(null);
-      showError('Webhook changed', 'Reload the registrations before trying this action again.');
-    }
-  }, [showError]);
-
-  const revealSecret = useCallback((
-    response: Pick<WebhookSecretResponse, 'signing_secret' | 'replayed'>,
-    operation: SecretOperation,
-  ) => {
-    const nextSecret = {
-      value: response.signing_secret,
-      replayed: response.replayed,
-      operation,
-    };
-    secretRef.current = nextSecret;
-    setSecretState(nextSecret);
-    setSecretCopied(false);
-    setSecretAcknowledged(false);
-    setSecretWarning('');
-  }, []);
-
-  const runSecretCommand = useCallback(async (
-    pending: PendingSecretCommand,
-    retry: boolean,
-  ) => {
-    setCommandBusy(true);
-    setCommandError('');
-    try {
-      const response = retry ? await pending.command.retry() : await pending.command.run();
-      pendingCommandRef.current = null;
-      setPendingOperation(null);
-      setCreateOpen(false);
-      clearCreateForm();
-      revealSecret(response.data, pending.operation);
-      success(
-        pending.operation === 'create' ? 'Webhook created' : 'Signing secret generated',
-        'The registration remains inactive until explicitly enabled.',
-      );
-      await loadControlPlane(pending.operation === 'create' ? 0 : offset);
-    } catch (error) {
-      if (pending.command.canRetry) {
-        setCommandError(
-          'The connection was lost after submission. Retry the same command, or reload and inspect inactive registrations before creating another one. A lost secret can only be replaced by generating a new secret.',
-        );
-      } else {
-        pendingCommandRef.current = null;
-        setPendingOperation(null);
-        if (pending.webhookId !== null && isConditionalError(error)) {
-          await recoverConditionalConflict(error, pending.webhookId, 'secret rotation');
-        } else {
-          const bounded = safeError(error, 'The webhook command failed.');
-          showError('Webhook command failed', bounded.message);
-        }
-      }
-    } finally {
-      setCommandBusy(false);
-    }
-  }, [clearCreateForm, loadControlPlane, offset, recoverConditionalConflict, revealSecret, showError, success]);
-
-  const beginCanonicalCreate = async () => {
-    if (!catalog || createEvents.length === 0 || !createUrl.trim()) return;
-    const timeout = Number(createTimeout);
-    if (!Number.isInteger(timeout) || timeout < 1 || timeout > 30) {
-      setCommandError('Timeout must be a whole number from 1 to 30 seconds.');
-      return;
-    }
-    const body: WebhookCreateRequest = {
-      url: createUrl.trim(),
-      event_types: [...createEvents],
-      description: createDescription.trim(),
-      timeout_seconds: timeout,
-    };
-    const command = createIdempotentCommand<WebhookCreateRequest, SecretCommandResult>(
-      'create',
-      body,
-      ({ body: requestBody, idempotencyKey }) => canonicalWebhookApi.createWebhook(
-        requestBody,
-        idempotencyKey,
-      ),
-    );
-    const pending = { command, operation: 'create' as const, webhookId: null };
-    pendingCommandRef.current = pending;
-    setPendingOperation('create');
-    setCreating(true);
-    try {
-      await runSecretCommand(pending, false);
-    } finally {
-      setCreating(false);
-    }
-  };
-
-  const retrySecretCommand = async () => {
-    const pending = pendingCommandRef.current;
-    if (!pending) return;
-    setCreating(pending.operation === 'create');
-    try {
-      await runSecretCommand(pending, true);
-    } finally {
-      setCreating(false);
-    }
-  };
-
-  const beginLegacyCreate = async () => {
-    const events = legacyEvents.split(',').map((event) => event.trim()).filter(Boolean);
-    if (!createUrl.trim() || events.length === 0) return;
-    setCreating(true);
-    setCommandError('');
-    try {
-      const response = await legacyWebhookApi.createWebhook({
-        url: createUrl.trim(),
-        events,
-        enabled: legacyEnabled,
-      });
-      setCreateOpen(false);
-      clearCreateForm();
-      revealSecret({ signing_secret: response.signingSecret, replayed: false }, 'create');
-      success('Legacy webhook created');
-      await loadControlPlane(0);
-    } catch {
-      showError('Webhook creation failed', 'The legacy webhook could not be created.');
-    } finally {
-      setCreating(false);
-    }
-  };
-
-  const openCreate = () => {
-    clearCreateForm();
-    setCommandError('');
-    setCreateOpen(true);
-  };
-
-  const handleCreateOpenChange = (open: boolean) => {
-    if (!open && pendingCommandRef.current) {
-      setCommandError('Resolve or reload the submitted command before closing this dialog.');
-      return;
-    }
-    setCreateOpen(open);
-    if (!open) clearCreateForm();
-  };
-
-  const toggleCreateEvent = (eventType: string) => {
-    setCreateEvents((current) => current.includes(eventType)
-      ? current.filter((entry) => entry !== eventType)
-      : [...current, eventType]);
-  };
-
-  const toggleEditEvent = (eventType: string) => {
-    setEditEvents((current) => current.includes(eventType)
-      ? current.filter((entry) => entry !== eventType)
-      : [...current, eventType]);
-  };
-
-  const openMetadataEditor = (registration: WebhookRegistration) => {
-    setEditor({ kind: 'metadata', registration });
-    setEditDescription(registration.description);
-    setEditTimeout(String(registration.timeout_seconds));
-    setEditEvents([...registration.event_types]);
-    setReplacementUrl('');
-    setConflict(null);
-  };
-
-  const openDestinationEditor = (registration: WebhookRegistration) => {
-    setEditor({ kind: 'destination', registration });
-    setReplacementUrl('');
-    setConflict(null);
-  };
-
-  const performConditionalUpdate = async (
-    webhookId: number,
-    body: WebhookPatchRequest,
-    action: string,
-  ) => {
-    setMutatingId(webhookId);
-    try {
-      const current = await canonicalWebhookApi.getWebhook(webhookId);
-      const confirmed = await promptPrivileged({
-        title: action,
-        message: registrationReviewSummary(current.data),
-        confirmText: action,
-        confirmationOnly: true,
-      });
-      if (!confirmed) return;
-      await canonicalWebhookApi.updateWebhook(webhookId, body, current.etag);
-      setEditor(null);
-      setConflict(null);
-      success('Webhook updated');
-      await loadControlPlane(offset);
-    } catch (error) {
-      setEditor(null);
-      if (isConditionalError(error)) {
-        await recoverConditionalConflict(error, webhookId, action);
-      } else {
-        const bounded = safeError(error, 'The webhook could not be updated.');
-        showError('Webhook update failed', bounded.message);
-      }
-    } finally {
-      setMutatingId(null);
-    }
-  };
-
-  const submitEditor = async () => {
-    if (!editor) return;
-    if (editor.kind === 'destination') {
-      if (!replacementUrl.trim()) return;
-      await performConditionalUpdate(
-        editor.registration.id,
-        { url: replacementUrl.trim() },
-        'Replace destination',
-      );
-      return;
-    }
-    const timeout = Number(editTimeout);
-    if (!Number.isInteger(timeout) || timeout < 1 || timeout > 30 || editEvents.length === 0) {
-      showError('Invalid webhook metadata', 'Select at least one event and use a timeout from 1 to 30 seconds.');
-      return;
-    }
-    await performConditionalUpdate(editor.registration.id, {
-      description: editDescription.trim(),
-      event_types: [...editEvents],
-      timeout_seconds: timeout,
-    }, 'Save changes');
-  };
-
-  const toggleCanonicalRegistration = async (registration: WebhookRegistration) => {
-    if (!status) return;
-    const blockReason = activationBlockReason(registration, status);
-    if (blockReason) return;
-    await performConditionalUpdate(
-      registration.id,
-      { active: !registration.active },
-      registration.active ? 'Disable webhook' : 'Enable webhook',
-    );
-  };
-
-  const deleteCanonicalRegistration = async (registration: WebhookRegistration) => {
-    setMutatingId(registration.id);
-    try {
-      const current = await canonicalWebhookApi.getWebhook(registration.id);
-      const confirmed = await promptPrivileged({
-        title: 'Delete webhook',
-        message: `${registrationReviewSummary(current.data)} Delete this registration? This cannot be undone.`,
-        confirmText: 'Delete webhook',
-        confirmationOnly: true,
-      });
-      if (!confirmed) return;
-      await canonicalWebhookApi.deleteWebhook(registration.id, current.etag);
-      success('Webhook deleted');
-      await loadControlPlane(0);
-    } catch (error) {
-      if (isConditionalError(error)) {
-        await recoverConditionalConflict(error, registration.id, 'delete');
-      } else {
-        const bounded = safeError(error, 'The webhook could not be deleted.');
-        showError('Webhook deletion failed', bounded.message);
-      }
-    } finally {
-      setMutatingId(null);
-    }
-  };
-
-  const rotateCanonicalSecret = async (registration: WebhookRegistration) => {
-    setMutatingId(registration.id);
-    try {
-      const current = await canonicalWebhookApi.getWebhook(registration.id);
-      if (current.data.active) {
-        showError('Rotation blocked', 'Disable the webhook before generating a new signing secret.');
-        return;
-      }
-      const confirmed = await promptPrivileged({
-        title: 'Generate a new signing secret',
-        message: `${registrationReviewSummary(current.data)} Generate a new signing secret? The previous secret will stop working.`,
-        confirmText: 'Generate secret',
-        confirmationOnly: true,
-      });
-      if (!confirmed) return;
-      const command = createIdempotentCommand<
-        { webhookId: number; etag: string },
-        SecretCommandResult
-      >(
-        'rotate',
-        { webhookId: registration.id, etag: current.etag },
-        ({ body, idempotencyKey }) => canonicalWebhookApi.rotateWebhookSecret(
-          body.webhookId,
-          body.etag,
-          idempotencyKey,
-        ),
-      );
-      const pending = {
-        command,
-        operation: 'rotate' as const,
-        webhookId: registration.id,
-      };
-      pendingCommandRef.current = pending;
-      setPendingOperation('rotate');
-      await runSecretCommand(pending, false);
-    } catch (error) {
-      if (isConditionalError(error)) {
-        await recoverConditionalConflict(error, registration.id, 'secret rotation');
-      } else {
-        const bounded = safeError(error, 'A new signing secret could not be generated.');
-        showError('Secret rotation failed', bounded.message);
-      }
-    } finally {
-      setMutatingId(null);
-    }
-  };
-
-  const handleCopySecret = async () => {
-    const current = secretRef.current;
-    if (!current) return;
-    try {
-      await navigator.clipboard.writeText(current.value);
-      setSecretCopied(true);
-      setSecretWarning('');
-    } catch {
-      setSecretWarning('Clipboard access failed. Select the secret and copy it manually.');
-    }
-  };
-
-  const requestSecretClose = () => {
-    if (!secretCopied || !secretAcknowledged) {
-      setSecretWarning('Copy and acknowledge the secret before closing this dialog.');
-      return;
-    }
-    clearSensitiveCommandState(false);
-  };
-
-  const toggleLegacyEnabled = async (registration: LegacyWebhookView) => {
-    try {
-      await legacyWebhookApi.updateWebhook(registration.id, { enabled: !registration.enabled });
-      success(registration.enabled ? 'Legacy webhook disabled' : 'Legacy webhook enabled');
-      await loadControlPlane(offset);
-    } catch {
-      showError('Legacy webhook update failed');
-    }
-  };
-
-  const deleteLegacyRegistration = async (registration: LegacyWebhookView) => {
-    const confirmed = await promptPrivileged({
-      title: 'Delete legacy webhook',
-      message: `Delete legacy webhook ${registration.id}? This cannot be undone.`,
-      confirmText: 'Delete webhook',
-      confirmationOnly: true,
-    });
-    if (!confirmed) return;
-    try {
-      await legacyWebhookApi.deleteWebhook(registration.id);
-      success('Legacy webhook deleted');
-      await loadControlPlane(0);
-    } catch {
-      showError('Legacy webhook deletion failed');
-    }
-  };
-
-  const testLegacyRegistration = async (registration: LegacyWebhookView) => {
-    try {
-      const delivery = await legacyWebhookApi.testWebhook(registration.id);
-      if (delivery.success) {
-        success('Legacy test delivery succeeded');
-      } else {
-        showError('Legacy test delivery failed');
-      }
-      if (legacyExpandedId === registration.id) {
-        const history = await legacyWebhookApi.getWebhookDeliveries(registration.id, {
-          limit: 50,
-          offset: 0,
-        });
-        setLegacyDeliveries(history.items);
-      }
-    } catch {
-      showError('Legacy test delivery failed');
-    }
-  };
-
-  const toggleLegacyDeliveries = async (registration: LegacyWebhookView) => {
-    if (legacyExpandedId === registration.id) {
-      setLegacyExpandedId(null);
-      setLegacyDeliveries([]);
-      return;
-    }
-    setLegacyExpandedId(registration.id);
-    setLegacyDeliveryLoading(true);
-    try {
-      const history = await legacyWebhookApi.getWebhookDeliveries(registration.id, {
-        limit: 50,
-        offset: 0,
-      });
-      setLegacyDeliveries(history.items);
-    } catch {
-      setLegacyDeliveries([]);
-      showError('Legacy delivery history could not be loaded');
-    } finally {
-      setLegacyDeliveryLoading(false);
-    }
-  };
-
-  const canonicalCreateBlocked = mode === 'canonical' && (
-    !status
-    || !canLoadCanonicalData(status)
-    || status.key_state !== 'available'
-    || status.limits.registrations_over_limit
-  );
-  const addDisabled = loading || mode === null || canonicalCreateBlocked;
-  const hasCanonicalNext = offset + canonicalPage.limit < canonicalPage.total;
-  const hasLegacyNext = offset + PAGE_SIZE < legacyTotal;
-  const visibleTotal = mode === 'canonical' ? canonicalPage.total : legacyTotal;
-  const visibleCount = mode === 'canonical' ? canonicalPage.items.length : legacyItems.length;
-  const hasPrevious = offset > 0;
-  const hasNext = mode === 'canonical' ? hasCanonicalNext : hasLegacyNext;
-
-  const goToPage = async (nextOffset: number) => {
-    const boundedOffset = Math.max(0, nextOffset);
-    if (mode === 'canonical') {
-      await loadCanonicalPage(boundedOffset);
-      return;
-    }
-    await loadControlPlane(boundedOffset);
-  };
+  const {
+    mode,
+    status,
+    catalog,
+    canonicalPage,
+    legacyItems,
+    offset,
+    loading,
+    statusError,
+    createOpen,
+    createUrl,
+    createDescription,
+    createTimeout,
+    createEvents,
+    legacyEvents,
+    legacyEnabled,
+    creating,
+    editor,
+    editDescription,
+    editTimeout,
+    editEvents,
+    replacementUrl,
+    replacementUrlError,
+    mutatingId,
+    conflict,
+    secretState,
+    secretCopied,
+    secretAcknowledged,
+    secretWarning,
+    commandError,
+    commandBusy,
+    pendingOperation,
+    hasPendingCommand,
+    legacyExpandedId,
+    legacyDeliveries,
+    legacyDeliveryLoading,
+    addDisabled,
+    visibleTotal,
+    visibleCount,
+    hasPrevious,
+    hasNext,
+    setCommandError,
+    setCreateDescription,
+    setCreateOpen,
+    setCreateTimeout,
+    setCreateUrl,
+    setEditDescription,
+    setEditTimeout,
+    setEditor,
+    setLegacyEnabled,
+    setLegacyEvents,
+    setReplacementUrl,
+    setReplacementUrlError,
+    setSecretAcknowledged,
+    clearSensitiveCommandState,
+    loadControlPlane,
+    retrySecretCommand,
+    beginCanonicalCreate,
+    beginLegacyCreate,
+    openCreate,
+    handleCreateOpenChange,
+    toggleCreateEvent,
+    toggleEditEvent,
+    openMetadataEditor,
+    openDestinationEditor,
+    submitEditor,
+    toggleCanonicalRegistration,
+    deleteCanonicalRegistration,
+    rotateCanonicalSecret,
+    handleCopySecret,
+    requestSecretClose,
+    toggleLegacyEnabled,
+    deleteLegacyRegistration,
+    testLegacyRegistration,
+    toggleLegacyDeliveries,
+    goToPage,
+  } = useWebhooksPageController();
 
   return (
     <ResponsiveLayout>
@@ -1184,7 +553,7 @@ function WebhooksPageContent() {
                 type="button"
                 variant="outline"
                 size="sm"
-                onClick={() => void goToPage(offset - PAGE_SIZE)}
+                onClick={() => void goToPage(offset - WEBHOOK_PAGE_SIZE)}
                 disabled={!hasPrevious || loading}
                 aria-label="Previous page"
               >
@@ -1195,7 +564,7 @@ function WebhooksPageContent() {
                 type="button"
                 variant="outline"
                 size="sm"
-                onClick={() => void goToPage(offset + PAGE_SIZE)}
+                onClick={() => void goToPage(offset + WEBHOOK_PAGE_SIZE)}
                 disabled={!hasNext || loading}
                 aria-label="Next page"
               >
@@ -1225,9 +594,14 @@ function WebhooksPageContent() {
                 type="url"
                 autoComplete="off"
                 value={createUrl}
-                onChange={(event) => setCreateUrl(event.target.value)}
+                onChange={(event) => {
+                  setCreateUrl(event.target.value);
+                  setCommandError('');
+                }}
                 placeholder="https://receiver.example/hooks/events"
-                disabled={Boolean(pendingCommandRef.current)}
+                maxLength={2_048}
+                aria-invalid={Boolean(commandError)}
+                disabled={hasPendingCommand}
               />
             </div>
             {mode === 'canonical' ? (
@@ -1239,7 +613,7 @@ function WebhooksPageContent() {
                     value={createDescription}
                     onChange={(event) => setCreateDescription(event.target.value)}
                     maxLength={500}
-                    disabled={Boolean(pendingCommandRef.current)}
+                    disabled={hasPendingCommand}
                   />
                 </div>
                 <div className="space-y-2">
@@ -1252,7 +626,7 @@ function WebhooksPageContent() {
                     step={1}
                     value={createTimeout}
                     onChange={(event) => setCreateTimeout(event.target.value)}
-                    disabled={Boolean(pendingCommandRef.current)}
+                    disabled={hasPendingCommand}
                   />
                 </div>
                 <fieldset className="space-y-3">
@@ -1262,7 +636,7 @@ function WebhooksPageContent() {
                       <Checkbox
                         checked={createEvents.includes(event.event_type)}
                         onCheckedChange={() => toggleCreateEvent(event.event_type)}
-                        disabled={Boolean(pendingCommandRef.current)}
+                        disabled={hasPendingCommand}
                       />
                       <span className="min-w-0">
                         <span className="block break-all font-mono text-sm">{event.event_type}</span>
@@ -1335,7 +709,7 @@ function WebhooksPageContent() {
               onClick={() => void (mode === 'canonical' ? beginCanonicalCreate() : beginLegacyCreate())}
               disabled={
                 creating
-                || Boolean(pendingCommandRef.current)
+                || hasPendingCommand
                 || !createUrl.trim()
                 || (mode === 'canonical' ? createEvents.length === 0 : !legacyEvents.trim())
               }
@@ -1368,9 +742,20 @@ function WebhooksPageContent() {
                 type="url"
                 autoComplete="off"
                 value={replacementUrl}
-                onChange={(event) => setReplacementUrl(event.target.value)}
+                onChange={(event) => {
+                  setReplacementUrl(event.target.value);
+                  setReplacementUrlError('');
+                }}
                 placeholder="https://receiver.example/hooks/new"
+                maxLength={2_048}
+                aria-invalid={Boolean(replacementUrlError)}
+                aria-describedby={replacementUrlError ? 'webhook-replacement-url-error' : undefined}
               />
+              {replacementUrlError && (
+                <p id="webhook-replacement-url-error" role="alert" className="text-sm text-destructive">
+                  {replacementUrlError}
+                </p>
+              )}
             </div>
           ) : editor ? (
             <div className="space-y-4">
