+++ admin-ui/app/webhooks/use-webhooks-page-controller.ts
@@ -1,0 +1,530 @@
+import { useCallback, useState } from 'react';
+
+import { usePrivilegedActionDialog } from '@/components/ui/privileged-action-dialog';
+import { useToast } from '@/components/ui/toast';
+import {
+  canonicalWebhookApi,
+  legacyWebhookApi,
+} from '@/lib/api-client';
+import type { LegacyWebhookView } from '@/lib/api-client';
+import { createIdempotentCommand } from '@/lib/idempotent-command';
+import type {
+  WebhookCreateRequest,
+  WebhookPatchRequest,
+  WebhookRegistration,
+} from '@/types';
+import {
+  activationBlockReason,
+  isConditionalWebhookError,
+  registrationReviewSummary,
+  safeWebhookError,
+  type PendingSecretCommand,
+  type SecretCommandResult,
+} from './webhook-controller-shared';
+import { useWebhookControlPlane } from './use-webhook-control-plane';
+import { useWebhookSecretCommands } from './use-webhook-secret-commands';
+import { validateWebhookUrl } from './webhook-url';
+
+export {
+  activationBlockReason,
+  canLoadCanonicalData,
+  WEBHOOK_PAGE_SIZE,
+} from './webhook-controller-shared';
+
+type EditorState = {
+  kind: 'metadata' | 'destination';
+  registration: WebhookRegistration;
+};
+
+export const useWebhooksPageController = () => {
+  const promptPrivileged = usePrivilegedActionDialog();
+  const { success, error: showError } = useToast();
+
+  const {
+    mode,
+    status,
+    catalog,
+    canonicalPage,
+    legacyItems,
+    offset,
+    loading,
+    statusError,
+    conflict,
+    legacyExpandedId,
+    legacyDeliveries,
+    legacyDeliveryLoading,
+    addDisabled,
+    visibleTotal,
+    visibleCount,
+    hasPrevious,
+    hasNext,
+    setConflict,
+    setLegacyExpandedId,
+    setLegacyDeliveries,
+    setLegacyDeliveryLoading,
+    loadControlPlane,
+    recoverConditionalConflict,
+    goToPage,
+  } = useWebhookControlPlane({ showError });
+
+  const [createOpen, setCreateOpen] = useState(false);
+  const [createUrl, setCreateUrl] = useState('');
+  const [createDescription, setCreateDescription] = useState('');
+  const [createTimeout, setCreateTimeout] = useState('10');
+  const [createEvents, setCreateEvents] = useState<string[]>([]);
+  const [legacyEvents, setLegacyEvents] = useState('');
+  const [legacyEnabled, setLegacyEnabled] = useState(true);
+  const [creating, setCreating] = useState(false);
+
+  const [editor, setEditor] = useState<EditorState | null>(null);
+  const [editDescription, setEditDescription] = useState('');
+  const [editTimeout, setEditTimeout] = useState('10');
+  const [editEvents, setEditEvents] = useState<string[]>([]);
+  const [replacementUrl, setReplacementUrl] = useState('');
+  const [replacementUrlError, setReplacementUrlError] = useState('');
+  const [mutatingId, setMutatingId] = useState<number | null>(null);
+
+  const clearCreateForm = useCallback(() => {
+    setCreateUrl('');
+    setCreateDescription('');
+    setCreateTimeout('10');
+    setCreateEvents([]);
+    setLegacyEvents('');
+    setLegacyEnabled(true);
+  }, []);
+
+  const {
+    secretState,
+    secretCopied,
+    secretAcknowledged,
+    secretWarning,
+    commandError,
+    commandBusy,
+    pendingOperation,
+    hasPendingCommand,
+    setCommandError,
+    setSecretAcknowledged,
+    clearSensitiveCommandState,
+    revealSecret,
+    startSecretCommand,
+    retrySecretCommand: retryPendingSecretCommand,
+    handleCopySecret,
+    requestSecretClose,
+  } = useWebhookSecretCommands({
+    clearCreateForm,
+    loadControlPlane,
+    offset,
+    recoverConditionalConflict,
+    setCreateOpen,
+    showError,
+    success,
+  });
+
+  const beginCanonicalCreate = async () => {
+    if (!catalog || createEvents.length === 0 || !createUrl.trim()) return;
+    const destination = validateWebhookUrl(createUrl);
+    if (!destination.valid) {
+      setCommandError(destination.message);
+      return;
+    }
+    const timeout = Number(createTimeout);
+    if (!Number.isInteger(timeout) || timeout < 1 || timeout > 30) {
+      setCommandError('Timeout must be a whole number from 1 to 30 seconds.');
+      return;
+    }
+    const body: WebhookCreateRequest = {
+      url: destination.value,
+      event_types: [...createEvents],
+      description: createDescription.trim(),
+      timeout_seconds: timeout,
+    };
+    const command = createIdempotentCommand<WebhookCreateRequest, SecretCommandResult>(
+      'create',
+      body,
+      ({ body: requestBody, idempotencyKey }) => canonicalWebhookApi.createWebhook(
+        requestBody,
+        idempotencyKey,
+      ),
+    );
+    const pending: PendingSecretCommand = { command, operation: 'create', webhookId: null };
+    setCreating(true);
+    try {
+      await startSecretCommand(pending);
+    } finally {
+      setCreating(false);
+    }
+  };
+
+  const retrySecretCommand = async () => {
+    if (!hasPendingCommand) return;
+    setCreating(pendingOperation === 'create');
+    try {
+      await retryPendingSecretCommand();
+    } finally {
+      setCreating(false);
+    }
+  };
+
+  const beginLegacyCreate = async () => {
+    const events = legacyEvents.split(',').map((event) => event.trim()).filter(Boolean);
+    if (!createUrl.trim() || events.length === 0) return;
+    const destination = validateWebhookUrl(createUrl);
+    if (!destination.valid) {
+      setCommandError(destination.message);
+      return;
+    }
+    setCreating(true);
+    setCommandError('');
+    try {
+      const response = await legacyWebhookApi.createWebhook({
+        url: destination.value,
+        events,
+        enabled: legacyEnabled,
+      });
+      setCreateOpen(false);
+      clearCreateForm();
+      revealSecret({ signing_secret: response.signingSecret, replayed: false }, 'create');
+      success('Legacy webhook created');
+      await loadControlPlane(0);
+    } catch {
+      showError('Webhook creation failed', 'The legacy webhook could not be created.');
+    } finally {
+      setCreating(false);
+    }
+  };
+
+  const openCreate = () => {
+    clearCreateForm();
+    setCommandError('');
+    setCreateOpen(true);
+  };
+
+  const handleCreateOpenChange = (open: boolean) => {
+    if (!open && hasPendingCommand) {
+      setCommandError('Resolve or reload the submitted command before closing this dialog.');
+      return;
+    }
+    setCreateOpen(open);
+    if (!open) clearCreateForm();
+  };
+
+  const toggleCreateEvent = (eventType: string) => {
+    setCreateEvents((current) => current.includes(eventType)
+      ? current.filter((entry) => entry !== eventType)
+      : [...current, eventType]);
+  };
+
+  const toggleEditEvent = (eventType: string) => {
+    setEditEvents((current) => current.includes(eventType)
+      ? current.filter((entry) => entry !== eventType)
+      : [...current, eventType]);
+  };
+
+  const openMetadataEditor = (registration: WebhookRegistration) => {
+    setEditor({ kind: 'metadata', registration });
+    setEditDescription(registration.description);
+    setEditTimeout(String(registration.timeout_seconds));
+    setEditEvents([...registration.event_types]);
+    setReplacementUrl('');
+    setReplacementUrlError('');
+    setConflict(null);
+  };
+
+  const openDestinationEditor = (registration: WebhookRegistration) => {
+    setEditor({ kind: 'destination', registration });
+    setReplacementUrl('');
+    setReplacementUrlError('');
+    setConflict(null);
+  };
+
+  const performConditionalUpdate = async (
+    webhookId: number,
+    body: WebhookPatchRequest,
+    action: string,
+  ) => {
+    setMutatingId(webhookId);
+    try {
+      const current = await canonicalWebhookApi.getWebhook(webhookId);
+      const confirmed = await promptPrivileged({
+        title: action,
+        message: registrationReviewSummary(current.data),
+        confirmText: action,
+        confirmationOnly: true,
+      });
+      if (!confirmed) return;
+      await canonicalWebhookApi.updateWebhook(webhookId, body, current.etag);
+      setEditor(null);
+      setConflict(null);
+      success('Webhook updated');
+      await loadControlPlane(offset);
+    } catch (error) {
+      setEditor(null);
+      if (isConditionalWebhookError(error)) {
+        await recoverConditionalConflict(error, webhookId, action);
+      } else {
+        const bounded = safeWebhookError(error, 'The webhook could not be updated.');
+        showError('Webhook update failed', bounded.message);
+      }
+    } finally {
+      setMutatingId(null);
+    }
+  };
+
+  const submitEditor = async () => {
+    if (!editor) return;
+    if (editor.kind === 'destination') {
+      if (!replacementUrl.trim()) return;
+      const destination = validateWebhookUrl(replacementUrl);
+      if (!destination.valid) {
+        setReplacementUrlError(destination.message);
+        return;
+      }
+      setReplacementUrlError('');
+      await performConditionalUpdate(
+        editor.registration.id,
+        { url: destination.value },
+        'Replace destination',
+      );
+      return;
+    }
+    const timeout = Number(editTimeout);
+    if (!Number.isInteger(timeout) || timeout < 1 || timeout > 30 || editEvents.length === 0) {
+      showError('Invalid webhook metadata', 'Select at least one event and use a timeout from 1 to 30 seconds.');
+      return;
+    }
+    await performConditionalUpdate(editor.registration.id, {
+      description: editDescription.trim(),
+      event_types: [...editEvents],
+      timeout_seconds: timeout,
+    }, 'Save changes');
+  };
+
+  const toggleCanonicalRegistration = async (registration: WebhookRegistration) => {
+    if (!status) return;
+    const blockReason = activationBlockReason(registration, status);
+    if (blockReason) return;
+    await performConditionalUpdate(
+      registration.id,
+      { active: !registration.active },
+      registration.active ? 'Disable webhook' : 'Enable webhook',
+    );
+  };
+
+  const deleteCanonicalRegistration = async (registration: WebhookRegistration) => {
+    setMutatingId(registration.id);
+    try {
+      const current = await canonicalWebhookApi.getWebhook(registration.id);
+      const confirmed = await promptPrivileged({
+        title: 'Delete webhook',
+        message: `${registrationReviewSummary(current.data)} Delete this registration? This cannot be undone.`,
+        confirmText: 'Delete webhook',
+        confirmationOnly: true,
+      });
+      if (!confirmed) return;
+      await canonicalWebhookApi.deleteWebhook(registration.id, current.etag);
+      success('Webhook deleted');
+      await loadControlPlane(0);
+    } catch (error) {
+      if (isConditionalWebhookError(error)) {
+        await recoverConditionalConflict(error, registration.id, 'delete');
+      } else {
+        const bounded = safeWebhookError(error, 'The webhook could not be deleted.');
+        showError('Webhook deletion failed', bounded.message);
+      }
+    } finally {
+      setMutatingId(null);
+    }
+  };
+
+  const rotateCanonicalSecret = async (registration: WebhookRegistration) => {
+    setMutatingId(registration.id);
+    try {
+      const current = await canonicalWebhookApi.getWebhook(registration.id);
+      if (current.data.active) {
+        showError('Rotation blocked', 'Disable the webhook before generating a new signing secret.');
+        return;
+      }
+      const confirmed = await promptPrivileged({
+        title: 'Generate a new signing secret',
+        message: `${registrationReviewSummary(current.data)} Generate a new signing secret? The previous secret will stop working.`,
+        confirmText: 'Generate secret',
+        confirmationOnly: true,
+      });
+      if (!confirmed) return;
+      const command = createIdempotentCommand<
+        { webhookId: number; etag: string },
+        SecretCommandResult
+      >(
+        'rotate',
+        { webhookId: registration.id, etag: current.etag },
+        ({ body, idempotencyKey }) => canonicalWebhookApi.rotateWebhookSecret(
+          body.webhookId,
+          body.etag,
+          idempotencyKey,
+        ),
+      );
+      const pending: PendingSecretCommand = {
+        command,
+        operation: 'rotate',
+        webhookId: registration.id,
+      };
+      await startSecretCommand(pending);
+    } catch (error) {
+      if (isConditionalWebhookError(error)) {
+        await recoverConditionalConflict(error, registration.id, 'secret rotation');
+      } else {
+        const bounded = safeWebhookError(error, 'A new signing secret could not be generated.');
+        showError('Secret rotation failed', bounded.message);
+      }
+    } finally {
+      setMutatingId(null);
+    }
+  };
+
+  const toggleLegacyEnabled = async (registration: LegacyWebhookView) => {
+    try {
+      await legacyWebhookApi.updateWebhook(registration.id, { enabled: !registration.enabled });
+      success(registration.enabled ? 'Legacy webhook disabled' : 'Legacy webhook enabled');
+      await loadControlPlane(offset);
+    } catch {
+      showError('Legacy webhook update failed');
+    }
+  };
+
+  const deleteLegacyRegistration = async (registration: LegacyWebhookView) => {
+    const confirmed = await promptPrivileged({
+      title: 'Delete legacy webhook',
+      message: `Delete legacy webhook ${registration.id}? This cannot be undone.`,
+      confirmText: 'Delete webhook',
+      confirmationOnly: true,
+    });
+    if (!confirmed) return;
+    try {
+      await legacyWebhookApi.deleteWebhook(registration.id);
+      success('Legacy webhook deleted');
+      await loadControlPlane(0);
+    } catch {
+      showError('Legacy webhook deletion failed');
+    }
+  };
+
+  const testLegacyRegistration = async (registration: LegacyWebhookView) => {
+    try {
+      const delivery = await legacyWebhookApi.testWebhook(registration.id);
+      if (delivery.success) {
+        success('Legacy test delivery succeeded');
+      } else {
+        showError('Legacy test delivery failed');
+      }
+      if (legacyExpandedId === registration.id) {
+        const history = await legacyWebhookApi.getWebhookDeliveries(registration.id, {
+          limit: 50,
+          offset: 0,
+        });
+        setLegacyDeliveries(history.items);
+      }
+    } catch {
+      showError('Legacy test delivery failed');
+    }
+  };
+
+  const toggleLegacyDeliveries = async (registration: LegacyWebhookView) => {
+    if (legacyExpandedId === registration.id) {
+      setLegacyExpandedId(null);
+      setLegacyDeliveries([]);
+      return;
+    }
+    setLegacyExpandedId(registration.id);
+    setLegacyDeliveryLoading(true);
+    try {
+      const history = await legacyWebhookApi.getWebhookDeliveries(registration.id, {
+        limit: 50,
+        offset: 0,
+      });
+      setLegacyDeliveries(history.items);
+    } catch {
+      setLegacyDeliveries([]);
+      showError('Legacy delivery history could not be loaded');
+    } finally {
+      setLegacyDeliveryLoading(false);
+    }
+  };
+
+  return {
+    mode,
+    status,
+    catalog,
+    canonicalPage,
+    legacyItems,
+    offset,
+    loading,
+    statusError,
+    createOpen,
+    createUrl,
+    createDescription,
+    createTimeout,
+    createEvents,
+    legacyEvents,
+    legacyEnabled,
+    creating,
+    editor,
+    editDescription,
+    editTimeout,
+    editEvents,
+    replacementUrl,
+    replacementUrlError,
+    mutatingId,
+    conflict,
+    secretState,
+    secretCopied,
+    secretAcknowledged,
+    secretWarning,
+    commandError,
+    commandBusy,
+    pendingOperation,
+    hasPendingCommand,
+    legacyExpandedId,
+    legacyDeliveries,
+    legacyDeliveryLoading,
+    addDisabled,
+    visibleTotal,
+    visibleCount,
+    hasPrevious,
+    hasNext,
+    setCommandError,
+    setCreateDescription,
+    setCreateOpen,
+    setCreateTimeout,
+    setCreateUrl,
+    setEditDescription,
+    setEditTimeout,
+    setEditor,
+    setLegacyEnabled,
+    setLegacyEvents,
+    setReplacementUrl,
+    setReplacementUrlError,
+    setSecretAcknowledged,
+    clearSensitiveCommandState,
+    loadControlPlane,
+    retrySecretCommand,
+    beginCanonicalCreate,
+    beginLegacyCreate,
+    openCreate,
+    handleCreateOpenChange,
+    toggleCreateEvent,
+    toggleEditEvent,
+    openMetadataEditor,
+    openDestinationEditor,
+    submitEditor,
+    toggleCanonicalRegistration,
+    deleteCanonicalRegistration,
+    rotateCanonicalSecret,
+    handleCopySecret,
+    requestSecretClose,
+    toggleLegacyEnabled,
+    deleteLegacyRegistration,
+    testLegacyRegistration,
+    toggleLegacyDeliveries,
+    goToPage,
+  };
+};
