+++ backlog/tasks/task-13014 - Implement-canonical-admin-webhook-control-plane-and-migration.md
@@ -4,7 +4,7 @@ title: Implement canonical admin webhook control plane and migration
 status: In Progress
 assignee: []
 created_date: '2026-08-21 20:41'
-updated_date: '2026-08-27 03:51'
+updated_date: '2026-08-27 04:45'
 labels:
   - admin
   - webhooks
@@ -198,6 +198,10 @@ The claimed nested-suite bug is factually incorrect and was not implemented. Rea
 2026-08-26 final rebase integration: after pushed source 51f204b5c1, origin/dev advanced 106 commits to 2306c1939f3b460f9c62da8ae83a1aa47c02ee0d and PR #2806 became conflicted. Rebased all 45 PR commits; the sole conflict was the generated OpenAPI fingerprint. Regenerated schema/types with the dependency-complete Python 3.11 environment: 2040 paths, 3037 schemas, sha256 a4c4b6b24aea9cfbaa4102595c29988e5ed8041e10574ff798773bf0d0ef09b4; drift check passes. Exact dev and rebased head both exposed two newly added moodboard-studio test files absent from every full-suite shard. Assigned both to chacha-content-persona in all five matrices. Their complete 95-case order also reproduced an exact-base StopIteration because a finite fake patched process-wide time.monotonic; production code was unchanged and the test now injects a module-local fake clock. Final shard result: 67 passed, 28 expected PostgreSQL skips. Rebased verification: 102 CI contracts passed, Admin_Webhooks non-PostgreSQL 302 passed/24 deselected, admin webhook UI 77 passed across 7 files, admin UI typecheck passed, focused Ruff/py_compile passed. Upstream frontend typecheck has 80 error lines byte-identical on exact dev and head, all in files unchanged by this PR; recorded as inherited base debt. Fresh Qodo and complete hosted CI at the rebased SHA remain mandatory.
 
 2026-08-26 final rebase contract reconciliation: a broader local gate added test_license_first_workflow_contracts.py to the prior 102-test ratchet/workflow subset and reproduced three stale freeze expectations from the intentional PR #2806 exact-base workflow changes. Updated the hand-checked frontend-required direct-trigger digest, direct pull-request base-reference count, and obsolete inline event-branch assertion; the contract now verifies the shared structured RATCHET_BASE_SHA expression on both frontend and admin ratchet steps. RED was 2 failures, then 1 deeper failure; final focused cases passed 2/2 and the complete CI ratchet/workflow set passed 116/116 with 2 warnings. All 22 embedded frontend-required Bash steps passed bash -n. Evidence updated in Docs/Evidence/Admin_Webhooks_PR1_Verification.md; exact-head hosted CI and Qodo remain mandatory.
+
+2026-08-26: User approved the bounded remediation design for the five active Qodo findings on PR #2806. Scope: explicit invariant URL validation for canonical/legacy create and destination replacement; extract page orchestration into focused control-plane/secret hooks plus a composing controller; exercise the real proxy response builder; add behavioral tests for the execution-order reporter; document the stdlib-only ratchet CLI logging exception and refresh pinned digests. Implementation will proceed test-first.
+
+2026-08-26 bounded Qodo remediation implementation is locally complete pending independent review. URL-validation TDD reproduced 3 unsafe create/replacement paths before implementation; proxy mutation proof reproduced the mocked-response gap (1 failed/15 passed). Added shared browser invariant validation, split the page into control-plane/secret hooks plus a composing controller, exercised the real proxy response builder, added direct execution-order reporter tests, and documented the stdlib-only ratchet status-output exception because frontend-unit-tests installs no Python dependencies before invoking the helper. Verification: focused admin UI 94/94; package typecheck; package lint 0 errors/41 inherited warnings; production build 49/49 routes (sandbox port-bind retry passed with normal process permissions); Chromium webhook lifecycle 1/1; CI/workflow contracts 129/129; all 22 embedded Bash steps; all pinned digests; OpenAPI drift; Admin_Webhooks non-PostgreSQL 302/302 with 24 PostgreSQL tests intentionally deselected; node/python syntax and git diff checks all pass. Independent review, immutable commit/evidence update, push, fresh Qodo, exact-head CI, and thread audit remain.
 <!-- SECTION:NOTES:END -->
 
 ## Definition of Done
